Added "DisableAutoSpaceChange" boolean to import
By default, a novice user will have a few flags set automatically
and a tranform prepended to the root. These changes convert from the
right-handed y-up world of Assimp to the left-handed z-up world of Unreal.

Should an advanced user prefer to control these changes themselves, the
"DisableAutoSpaceChange" option can revert to as-is Assimp importing.